Refactored (read the whole comment).
Now the application uses the public api service to process an image (or
just to obtain the representation of the previously processed image).
The public api either instantiates an image api or uses the previously
instantiated one, depending on the parameter of the processing request.
IMPORTANT: there are still methods that are never used and checks that
are being made repeatedly in different classes for different (or not)
purposes (too sleepy now to figure those out).